shrunk the RefOrValue interface to only include methods that are actually used per comments
